Results are now back in thread 0.

+++ src/cpp/sky-state-controller/SkyState.h
@@ -5,10 +5,10 @@
 
 class SkyState{
 public:
-  SkyState(AstroTime* astroTimePnt, Location* locationPnt, SkyManager* skyManagerPnt, double* memoryPtr);
+  SkyState(AstroTime* astroTimePnt, Location* locationPnt, SkyManager* skyManagerPnt, float* memoryPtr);
   Location* location;
   AstroTime* astroTime;
   SkyManager* skyManager;
-  double* memoryPtr;
+  float* memoryPtr;
   void updateHeap32Memory();
 };
